Added: command line option to force close cyberfox during install.
In rare cases some users are unable to install a update due to cyberfox
process is running, Some users find it hard to follow the documentation
on how to terminate the process so we have made it even easier for when
they encounter this non-common issue.

By passing --FORCE command line option to the installer it will
automatically kill any cyberfox.exe processes running before installing,
This will make it easier by way of a simple run command then explaining
how to use task manager.