Updated: Installer script to version 2.4.0
If distribution folder exists, When removing it leave {app} directory to
preserve user add files. i.e addons

If a user installs with create desktop shortcut selected then installs
again this time leaving the create desktop shortcut deselected, During
the uninstall process the desktop shortcut is not removed, Now the
desktop icon is removed.

If a user wishes to have a modified desktop shortcut all the user must
do is change the name of the shortcut.